% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{quantiles}
\alias{quantiles}
\title{Calculate Quantiles}
\usage{
quantiles(df, col, prob)
}
\arguments{
\item{df}{Data frame to use the function on}

\item{col}{Column name to perform calculations on}

\item{prob}{Numeric vector of probabilities to specify which quantiles to calculate}
}
\value{
A numeric vector of quantiles
}
\description{
Calculate Quantiles
}
\examples{
quantiles(mtcars, "mpg", probs = c(0.25, 0.5, 0.75))
}
