% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{statsparam}
\alias{statsparam}
\title{Calculate Mean, Median, Variance, and Standard Deviation}
\usage{
statsparam(df, col)
}
\arguments{
\item{df}{Data frame to use the function on}

\item{col}{Column name to perform calculations on}
}
\value{
A list with the computed mean, median, variance, and standard deviation
}
\description{
Calculate Mean, Median, Variance, and Standard Deviation
}
\examples{
stats(mtcars, "mpg")
}
